Fixing ktlint errors, and reorganizing code for the 'About' settings.
